refactor: session resource
