+++ apps/server/src/api-data/session/session.router.ts
@@ -1,0 +1,7 @@
+import express from 'express';
+
+import { getInfo } from './session.controller.js';
+
+export const router = express.Router();
+
+router.get('/info', getInfo);
