+++ apps/server/src/services/project-service/ProjectService.ts
@@ -1,11 +1,10 @@
-import { DatabaseModel, GetInfo, LogOrigin, ProjectData, ProjectFileListResponse } from 'ontime-types';
+import { DatabaseModel, LogOrigin, ProjectData, ProjectFileListResponse } from 'ontime-types';
 import { getErrorMessage } from 'ontime-utils';
 
 import { copyFile, rename } from 'fs/promises';
 
 import { logger } from '../../classes/Logger.js';
-import { getNetworkInterfaces } from '../../utils/networkInterfaces.js';
-import { publicDir, publicFiles } from '../../setup/index.js';
+import { publicDir } from '../../setup/index.js';
 import {
   appendToName,
   ensureDirectory,
@@ -304,26 +303,6 @@ export async function deleteProjectFile(filename: string) {
   }
 
   await deleteFile(projectFilePath);
-}
-
-/**
- * Adds business logic to gathering data for the info endpoint
- */
-export async function getInfo(): Promise<GetInfo> {
-  const { version, serverPort } = getDataProvider().getSettings();
-  const osc = getDataProvider().getOsc();
-
-  // get nif and inject localhost
-  const ni = getNetworkInterfaces();
-  ni.unshift({ name: 'localhost', address: '127.0.0.1' });
-
-  return {
-    networkInterfaces: ni,
-    version,
-    serverPort,
-    osc,
-    cssOverride: publicFiles.cssOverride,
-  };
 }
 
 /**
